packages/load_backup/pyproject.toml
[tool.poetry]
name = "load-backup"
version = "0.1.0"
description = ""
authors = ["Ethan Leifer <ethanleifer@gmail.com>"]
packages = [{include = "load_backup"}]

[tool.poetry.dependencies]
python = "^3.10"
gcloud-aio-storage = "^8.2.0"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
